fix(db): enable sqlx compile-time query checking (Block 3.3)
Convert all ~80 queries across 10 repository files from runtime
sqlx::query()/query_as() to compile-time checked query!()/query_as!()
macros. This catches SQL errors at build time rather than runtime.

- Fix SQL injection in BookRepository::list (format! interpolation →
  parameterized IS NULL OR filters with bind parameters)
- Tags filter uses json_each(?) with JSON array bind
- Dynamic ORDER BY via match-arm enumeration (20 sort variants for books)
- FTS search via JOIN variant; optional filters via IS NULL OR
- Add sqlx "macros" feature to workspace dependencies
- Generate and commit .sqlx/ for CI offline mode (SQLX_OFFLINE=true)
- pool.rs PRAGMA queries remain runtime (non-standard SQL)

+++ crates/archivis-api/src/books/types.rs
@@ -268,13 +268,14 @@ impl From<Book> for BookSummary {
 }
 
 impl From<BookAuthorEntry> for AuthorEntry {
+    #[allow(clippy::cast_possible_truncation)]
     fn from(entry: BookAuthorEntry) -> Self {
         Self {
             id: entry.author.id,
             name: entry.author.name,
             sort_name: entry.author.sort_name,
             role: entry.role,
-            position: entry.position,
+            position: entry.position as i32,
         }
     }
 }
